+++ test/authorization-code.ts
@@ -95,10 +95,7 @@ describe('authorization-code', () => {
           extraParams
         });
       } catch (error: any) {
-        expect(error.message).to.equal(
-          'The following extraParams are disallowed: \'client_id\', \'response_type\', \'redirect_uri\', ' +
-             '\'code_challenge_method\', \'code_challenge\', \'state\', \'scope\''
-        );
+        expect(error.message).to.include('Property in extraParams');
         return;
       }
 
@@ -128,6 +125,32 @@ describe('authorization-code', () => {
         await client.authorizationCode.getAuthorizeUri({
           redirectUri,
           codeVerifier,
+        })
+      ).to.equal(server.url + '/authorize?' + params.toString());
+
+    });
+    it('should support the resource parameter', async() => {
+
+      const server = testServer();
+      const client = new OAuth2Client({
+        server: server.url,
+        authorizationEndpoint: '/authorize',
+        clientId: 'test-client-id',
+      });
+
+      const redirectUri = 'http://my-app.example/redirect';
+      const resource = ['https://example/foo1', 'https://example/foo2'];
+      const params = new URLSearchParams({
+        client_id: 'test-client-id',
+        response_type: 'code',
+        redirect_uri: redirectUri,
+      });
+      for(const r of resource) params.append('resource', r);
+
+      expect(
+        await client.authorizationCode.getAuthorizeUri({
+          redirectUri,
+          resource,
         })
       ).to.equal(server.url + '/authorize?' + params.toString());
 
@@ -235,6 +258,74 @@ describe('authorization-code', () => {
       });
 
     });
+    it('should not use Basic Auth if no secret is provided, even if client_secret_basic is set.', async() => {
+
+      const server = testServer();
+
+      const client = new OAuth2Client({
+        server: server.url,
+        tokenEndpoint: '/token',
+        clientId: 'test-client-id',
+        authenticationMethod: 'client_secret_basic',
+      });
+
+      const result = await client.authorizationCode.getToken({
+        code: 'code_000',
+        redirectUri: 'http://example/redirect',
+      });
+
+      expect(result.accessToken).to.equal('access_token_000');
+      expect(result.refreshToken).to.equal('refresh_token_000');
+      expect(result.expiresAt).to.be.lessThanOrEqual(Date.now() + 3600_000);
+      expect(result.expiresAt).to.be.greaterThanOrEqual(Date.now() + 3500_000);
+
+      const request = server.lastRequest();
+      expect(request.headers.get('Authorization')).to.equal(null);
+
+      expect(request.body).to.eql({
+        client_id: 'test-client-id',
+        grant_type: 'authorization_code',
+        code: 'code_000',
+        redirect_uri: 'http://example/redirect',
+      });
+
+    });
+
+    it('should support the resource parameter', async() => {
+
+      const server = testServer();
+
+      const client = new OAuth2Client({
+        server: server.url,
+        tokenEndpoint: '/token',
+        clientId: 'test-client-id',
+      });
+      const resource = ['https://example/foo1', 'https://example/foo2'];
+
+      const result = await client.authorizationCode.getToken({
+        code: 'code_000',
+        redirectUri: 'http://example/redirect',
+        resource,
+      });
+
+      expect(result.accessToken).to.equal('access_token_000');
+      expect(result.refreshToken).to.equal('refresh_token_000');
+      expect(result.expiresAt).to.be.lessThanOrEqual(Date.now() + 3600_000);
+      expect(result.expiresAt).to.be.greaterThanOrEqual(Date.now() + 3500_000);
+
+      const request = server.lastRequest();
+      expect(request.headers.get('Authorization')).to.equal(null);
+
+      expect(request.body).to.eql({
+        client_id: 'test-client-id',
+        grant_type: 'authorization_code',
+        code: 'code_000',
+        redirect_uri: 'http://example/redirect',
+        resource,
+      });
+
+    });
+
 
   });
 });